% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1regularization.R
\name{LinearModelL1}
\alias{LinearModelL1}
\title{LinearModelL1}
\usage{
LinearModelL1(X.scaled.mat, y.vec, penalty, opt.thresh, initial.weight.vec,
  step.size)
}
\arguments{
\item{X.scaled.mat}{scaled feature matrix (n_observations x n_features)}

\item{y.vec}{label vector (n_observations x 1)}

\item{penalty}{non-negative numeric scalar}

\item{opt.thresh}{positive numeric scalar, threshold on the sub-differential optimality criterion}

\item{initial.weight.vec}{weight vector (n_observations x 1)}

\item{step.size}{int scalar > 1}
}
\value{
w.opt optimal weight vector (n_observations+1 x 1)
}
\description{
R function that calculates the L1 Regularization
}
\examples{

}
