% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1regularization.R
\name{LinearModelL1CV}
\alias{LinearModelL1CV}
\title{LinearModelL1CV}
\usage{
LinearModelL1CV(X.mat, y.vec, fold.vec, n.folds = 5, penalty.vec,
  step.size)
}
\arguments{
\item{X.mat}{feature matrix (n_observations x n_features)}

\item{y.vec}{label vector (n_observations x 1)}

\item{fold.vec}{fold ID vector (n_observations x 1)}

\item{n.folds}{int scalar > 1}

\item{step.size}{int scalar > 1}
}
\value{
mean.validation.loss, mean.train.loss.vec, penalty.vec, selected.penalty, weight.vec, predict=function(testX.mat)
}
\description{
R function that calculates the L1 cross-validation
}
